+++ core/nvim/lua/gerrrt/plugins/mini-nvim.lua
@@ -66,7 +66,6 @@ return {
 			},
 		})
 		require("mini.cursorword").setup({})
-		require("mini.indentscope").setup({})
 		require("mini.pairs").setup({})
 		require("mini.trailspace").setup({})
 		require("mini.bufremove").setup({})